refactor: destroyOmbul
